# src/main/resources/META-INF/openapi.yaml
openapi: 3.0.0
info:
    description: "This is the REST 'contract' for the Order System"
    version: 1.0.0
    title: "(20F) Group Project - Order System"
    contact:
        name: "(20F) CST 8277"
        email: "normanm@algonquincollege.com"
        url: "https://brightspace.algonquincollege.com/d2l/home/279461"
servers:
    -
        url: "http://localhost:{port}/{contextRoot}/{basePath}"
        variables:
            port: {
                default: "8080"
            }
            contextRoot: {
                default: "rest-orderSystem"
            }
            basePath: {
                default: "api/v1"
            }
security:
    -
        basicAuth: []
...
components:
    securitySchemes:
        basicAuth:
            type: http
            scheme: basic
    parameters:
        idPathParam:
            in: path
            name: id
            required: true
            schema: {
                type: integer,
                format: int64,
                description: "a resource's id",
                example: 1
            }
    responses:
        ForbiddenError:
            description: "Access to the specified resource has been forbidden"
        UnauthorizedAdminError:
            description: "request requires HTTP authentication for ADMIN_ROLE"
        UnauthorizedUserError:
            description: "request requires HTTP authentication for USER_ROLE"
        NotFound:
            description: "The specified resource was not found"
    schemas:
        PojoBase:
            type: object
            required: [id, version]
            properties: {
                id: {
                    description: "id of entity",
                    type: integer
                },
                version: {
                    description: "version (OptLock) of entity",
                    type: integer
                },
                createdDate: {
                    description: "creation timestamp of entity",
                    type: string,
                    format: date-time
                    },
                updatedDate: {
                    description: "last-updatd timestamp of entity",
                    type: string,
                    format: date-time
                }
            }
...
        Order:
            description: "(20F) Group Project - Order entity"
            type: object
            allOf: [{
                $ref: "#/components/schemas/PojoBase"
            }]
            properties: {
                description: {type: string, description: "Order description", example: My order}
            }
...
        OrderLine:
            description: "(20F) Group Project - Order Line entity"
            type: object
            properties: {
                orderId: {type: integer, description: "Order Id", example: 1},
                amount: {type: number, description: "Amount"},
                productId: {type: integer, description: "Product Id", example: 2}
            }
        OrderLinesList:
            description: "(20F) Group Project - Order Lines List entity"
            type: object
            properties: {
                orderId: {type: integer, description: "Order Id", example: 1},
                amount: {type: number, description: "Amount"},
                productId: {type: integer, description: "Product Id", example: 2}
            }
paths:
...
    /order/{id}:
        get:
            summary: "Retrieves an order by its id"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            }]
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedUserError"
                },
                "403": {
                    $ref: "#/components/responses/ForbiddenError"
                },
                "404": {
                    $ref: "#/components/responses/NotFound"
                }
            }
            tags:
               - orders
            
        delete:
            summary: "Removes an order by its id"
            parameters: [{
                $ref: "#/components/parameters/idPathParam"
            }]
            responses: {
                "204": {
                    description: "Ok (No content, successful operation)"
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedUserError"
                },
                "403": {
                    $ref: "#/components/responses/ForbiddenError"
                },
                "404": {
                    $ref: "#/components/responses/NotFound"
                }
            }
            tags:
               - orders
    /order/orderLines/{orderId}:
        get:
            summary: "Get order lines by order Id"
            parameters:
                - name: "orderId"
                  in: "path"
                  description: "Order Id"
                  required: true
                  type: "integer"
                  format: "int64"
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/OrderLinesList"
                            }
                        }
                    }
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedAdminError"
                }
            }
            tags:
               - orders
    /order/orderLine/{orderId}/{orderLineNumber}:
        get:
            summary: "Retrieves an order line by its id"
            parameters:
                - name: "orderId"
                  in: "path"
                  description: "Order Id"
                  required: true
                  type: "integer"
                  format: "int64"
                - name: "orderLineNumber"
                  in: "path"
                  description: "Order Line Number"
                  required: true
                  type: "integer"
                  format: "int64"
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/OrderLine"
                            }
                        }
                    }
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedUserError"
                },
                "403": {
                    $ref: "#/components/responses/ForbiddenError"
                },
                "404": {
                    $ref: "#/components/responses/NotFound"
                }
            }
            tags:
               - orders
    /order/addOrderLineToOder/{productId}/{orderId}/{amount}:
       put:
            summary: "Add order line to order"
            parameters:
                - name: "productId"
                  in: "path"
                  description: "Product Id"
                  required: true
                  type: "integer"
                  format: "int64"
                - name: "orderId"
                  in: "path"
                  description: "Order Id"
                  required: true
                  type: "integer"
                  format: "int64"
                - name: "amount"
                  in: "path"
                  description: "Amount"
                  required: true
                  type: "number"
                  format: "float"
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/OrderLine"
                            }
                        }
                    }
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedUserError"
                },
                "403": {
                    $ref: "#/components/responses/ForbiddenError"
                },
                "404": {
                    $ref: "#/components/responses/NotFound"
                }
            }
            tags:
               - orders
    /order/deleteOrderLineToOder/{orderLineNumber}/{orderId}:
       delete:
            summary: "Delete order line from order"
            parameters:
                - name: "orderLineNumber"
                  in: "path"
                  description: "Order Line Number"
                  required: true
                  type: "integer"
                  format: "int64"
                - name: "orderId"
                  in: "path"
                  description: "Order Id"
                  required: true
                  type: "integer"
                  format: "int64"
            responses: {
                "204": {
                    description: "Ok (No content, successful operation)"
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedUserError"
                },
                "403": {
                    $ref: "#/components/responses/ForbiddenError"
                },
                "404": {
                    $ref: "#/components/responses/NotFound"
                }
            }
            tags:
               - orders
    /order/updateOrderLineToOder/{orderLineNumber}/{orderId}/{amount}:
       put:
            summary: "Update order line"
            parameters:
                - name: "orderLineNumber"
                  in: "path"
                  description: "Order Line Number"
                  required: true
                  type: "integer"
                  format: "int64"
                - name: "orderId"
                  in: "path"
                  description: "Order Id"
                  required: true
                  type: "integer"
                  format: "int64"
                - name: "amount"
                  in: "path"
                  description: "Amount"
                  required: true
                  type: "number"
                  format: "float"
            responses: {
                "200": {
                    description: "Ok (Successful operation)",
                    content: {
                        application/json: {
                            schema: {
                                $ref: "#/components/schemas/OrderLine"
                            }
                        }
                    }
                },
                "401": {
                    $ref: "#/components/responses/UnauthorizedUserError"
                },
                "403": {
                    $ref: "#/components/responses/ForbiddenError"
                },
                "404": {
                    $ref: "#/components/responses/NotFound"
                }
            }
            tags:
               - orders     